refactor(top-bar): move all relative logic to `TopBar.vue` && remove `alwaysShowTheTopBarLogoOnSearchPageMode` setting
