+++ src/_locales/cmn-CN.yml
@@ -151,7 +151,6 @@ settings:
   home_tabs_adjustment: 标签栏调整
   home_tabs_adjustment_desc: 第一个激活的标签项是默认界面
   always_show_tabs_on_home_page: 总是显示首页标签栏
-  always_show_the_top_bar_logo: 总是显示顶栏 Logo
 
   # Compatibility
   # Common
