+++ src/_locales/en.yml
@@ -152,7 +152,6 @@ settings:
   home_tabs_adjustment: Tabs adjustment
   home_tabs_adjustment_desc: The first activated tab is the default page
   always_show_tabs_on_home_page: Always show tabs on the home page
-  always_show_the_top_bar_logo: Always show the top bar logo
 
   # Compatibility
   # Common
